Final optimization of CPU & time consuming, not compressing
I'm sorry, because i've needed to refactor my 2 previous commits. But finally it save almost 10 times the time needed to make a backup of about 7Gb of space. The solution was not only not merge the DB backup with the backup of the Home directory, but also NOT COMPRESS that Home directory copy. So instead of get a compressed TAR.GZ of the Home Directory we get a TAR file containing the Home Directory (or the files and subdirectories you specify in the profile).

Furthermore, when you specify not to compress i choosed to not to join both copies (DB + HOME DIR) and leave it as 2 different files per daily backup. At least for me has another advantage: i can get a copy of the database without decompress the daily backup.

My change introduced in this commit needs this in the profile (i will commit it in the next commit):

# Optional: if your files & directories to be backed up are too bigger in size, you can set this to false
# and then that backup will not be compressed and the the daily backup were composed by 2 files (SQL.GZ for DB and TAR for files&dirs)
compressed_tar_file=false

+++ WordPressBackup.sh
@@ -67,7 +67,7 @@ for backupprofile in $profile ; do
 
 			# Which files & sub-directories to backup of the root directory
 			if [ "${file_list}" = "" ]; then
-				file_list_bckp =  ${wp_root}
+				file_list_bckp=${wp_root}
 			else
 				files=(${file_list})
 				file_list_absolute=""
@@ -85,23 +85,44 @@ for backupprofile in $profile ; do
 			cd ${backup_location}/${user}/${wp_domain}
 
 			# MySQL Takes a Dump and compress the Home Directory
-			mysqldump -u ${db_user} --host ${db_host}  -p${db_pass} ${db_name} | gzip > ./${backupname}-DB.sql.gz &&
-			tar zcPf ./${backupname}-FILES.tar.gz ${file_list_bckp}
+			if [ "${compressed_tar_file}" != false ]; then
+				mysqldump -u ${db_user} --host ${db_host}  -p${db_pass} ${db_name} | gzip > ./${backupname}-DB.sql.gz &&
+				tar zcPf ./${backupname}-FILES.tar.gz ${file_list_bckp}
+				
+				# Compresses the MySQL Dump and the Home Directory
+				tar zcPf ./${wp_domain}-${backupname}.tar.gz ./${backupname}-FILES.tar.gz ./${backupname}-DB.sql.gz
+				chmod 600 ./${wp_domain}-${backupname}.tar.gz
 
-			# Compresses the MySQL Dump and the Home Directory
-			tar zcPf ./${wp_domain}-${backupname}.tar.gz ./${backupname}-FILES.tar.gz ./${backupname}-DB.sql.gz
-			chmod 600 ./${wp_domain}-${backupname}.tar.gz
+				# Generates the Backup Size
+				#FILENAME=${backup_location}/${user}/${wp_domain}/${wp_domain}-${backupname}.tar.gz
+				FILENAME=${wp_domain}-${backupname}.tar.gz
+				FILESIZE=$(du -h "$FILENAME")
+				if [ "${quiet}" = "0" ]; then
+					echo "$FILESIZE"
+				fi
 
-			# Generates the Backup Size
-			FILENAME=${backup_location}/${user}/${wp_domain}/${wp_domain}-${backupname}.tar.gz
-			FILESIZE=$(du -h "$FILENAME")
-			if [ "${quiet}" = "0" ]; then
-				echo "$FILESIZE"
+				#Removes the SQL dump and Home DIR to conserve space
+				rm -rf ./${backupname}-FILES.tar.gz ./${backupname}-DB.sql.gz
+
+			else
+				mysqldump -u ${db_user} --host ${db_host}  -p${db_pass} ${db_name} | gzip > ./${backupname}-DB.sql.gz &&
+				tar -cPf ./${backupname}-FILES.tar ${file_list_bckp}
+
+				# Generates the Backup files Size
+				if [ "${quiet}" = "0" ]; then
+					#FILENAME=${backup_location}/${user}/${wp_domain}/${backupname}-DB.sql.gz
+					FILENAME=${backupname}-DB.sql.gz
+					FILESIZE=$(du -h "$FILENAME")
+					echo "$FILESIZE"
+
+					#FILENAME=${backup_location}/${user}/${wp_domain}/${backupname}-FILES.tar
+					FILENAME=${backupname}-FILES.tar
+					FILESIZE=$(du -h "$FILENAME")
+					echo "$FILESIZE"
+				fi
+
 			fi
-
-			#Removes the SQL dump and Home DIR to conserve space
-			rm -rf ./${backupname}-FILES.tar.gz ./${backupname}-DB.sql.gz
-
+			
 			#Deletes any Backup older than X days
 			find ${backup_location}/${user}/${wp_domain}/ -type f -mtime +${keepdays} -exec rm {} \;
 		fi
